Support both zarr2/tifffile<2025.5.21 and zarr3/tifffile>2025.5.21 (#48)
Right now reading multiscale ndpi/svs is broken on py311+ because the
implementation here does not work with the upstream changes in tifffile.

This is a replacement for PR pinning back tifffile, to work with zarr<3
(https://github.com/napari/napari-tiff/pull/47), or one pinning up
tifffile, to work with zarr>3. Originally, I was going to do the latter
here (hence branch name), but over in the tifffile repo, cgohlke
suggested using the attrs approach that works with both the zarr2 stores
and zarr3 stores (the later lack the various list methods, so you get
errors if you try to use `.arrays()` as before).
See:
https://github.com/cgohlke/tifffile/issues/297#issuecomment-2906373547

I've tested this locally with a few ndpi as well as run tests--it works
with both combinations.
Importantly, it works with py310, which isn't supported by zarr3, but is
supported by napari.
In the issue linked above, there is an indication of performance
regression with zarr3 and it does, subjectively, feel worse. But
hopefully that is something that can be improved upstream.

In pyproject.toml I let tifffile sort out imagecodec and zarr
versioning, which does move us up a bit in tifffile version, but as
noted above, soon this will be moot as support for py310 is being
dropped more broadly.

Edit: the test suite here includes py310 and py312 so we should get both
combinations tested on CI.