ENH: for multiscale, use dask and auto-rechunk (#45)
Currently the plugin uses the zarr tifffile backend to read multiscale
tiff-variants, which includes ndpi, svs, and other whole-slide image
formats. Note: this is only supported with zarr<3)

In this PR, I wrap the arrays with dask, which has two primary
advantages:
1. it allows for re-chunking to happen ~~(`chunks='auto'`)~~ `chunks='1
MiB'` with a 1 MiB chunk mass, which is a major performance boost for
ndpi images from Hamamatsu slide scanners which have native chunks of
(8, 3840, 3). You can download some samples at:
  https://openslide.cs.cmu.edu/download/openslide-testdata/Hamamatsu/
  (I've also verified this with scans produced by JAX microscopy core.) 
For a svs (native 256, 256, 3) `chunks='1 MiB'` yields up as 512, 512, 3
and is nice and performant.
The ndpis end up as either chunksize=(32, 8192, 3) or chunksize=(24,
11520, 3) and are also much improved.
Nota bene: dask by default aims for 128 MiB chunks -- we may consider
making this configurable in napari:
https://github.com/napari/napari/issues/7924 -- which is not great for
visualization, hence specifying 1 MiB
2. by using dask, we can take advantage of caching, which increases
memory usage, but also contributes to improved performance.

Note: With this PR for the big ndpi i get more similar, though still
slightly worse performance than with QuPath. There is a stutter at the
resolution levels (without ASYNC) and with ASYNC it takes a split second
to load the 0 level.
If I manually set the chunks, e.g. ` chunks=(2048, 2048, 3))` or `
chunks=(1024, 1024, 3))` then performance is bit better for the npdi and
with ASYNC pretty much identical to QuPath.
But I think the auto-chunks are fine for the general use case.